Add Bot API routes extension
